Split routes into multiple files
